GC-463 - incrementing version to 0.0.10   + renaming migration script with new prefix because of collision with dorado-modeling   + reverted the un-wanted ModelMatrix class rename issue (thanks intelliJ)
