+++ modelmatrix-cli/src/main/scala/com/collective/modelmatrix/cli/definition/ViewFeatures.scala
@@ -1,6 +1,6 @@
 package com.collective.modelmatrix.cli.definition
 
-import com.collective.modelmatrix.ModelMatrixAccess.ModelMatrixCatalogAccess
+import com.collective.modelmatrix.ModelMatrix.ModelMatrixCatalogAccess
 import com.collective.modelmatrix.cli.Script
 import org.slf4j.LoggerFactory
