+++ modelmatrix-cli/src/main/scala/com/collective/modelmatrix/cli/instance/AddInstance.scala
@@ -1,6 +1,6 @@
 package com.collective.modelmatrix.cli.instance
 
-import com.collective.modelmatrix.ModelMatrixAccess.ModelMatrixCatalogAccess
+import com.collective.modelmatrix.ModelMatrix.ModelMatrixCatalogAccess
 import com.collective.modelmatrix._
 import com.collective.modelmatrix.cli.{SourceTransformation, CliSparkContext, Script, Source}
 import com.collective.modelmatrix.transform._
@@ -26,7 +26,7 @@ case class AddInstance(
 
   private val log = LoggerFactory.getLogger(classOf[AddInstance])
 
-  private implicit lazy val sqlContext = ModelMatrixAccess.hiveContext(sc)
+  private implicit lazy val sqlContext = ModelMatrix.hiveContext(sc)
 
   import com.collective.modelmatrix.cli.ASCIITableFormat._
   import com.collective.modelmatrix.cli.ASCIITableFormats._
