+++ modelmatrix-cli/src/main/scala/com/collective/modelmatrix/cli/instance/ValidateInputData.scala
@@ -1,7 +1,7 @@
 package com.collective.modelmatrix.cli.instance
 
-import com.collective.modelmatrix.ModelMatrixAccess
-import com.collective.modelmatrix.ModelMatrixAccess.ModelMatrixCatalogAccess
+import com.collective.modelmatrix.ModelMatrix
+import com.collective.modelmatrix.ModelMatrix.ModelMatrixCatalogAccess
 import com.collective.modelmatrix.cli.{Source, _}
 import com.collective.modelmatrix.transform._
 import org.slf4j.LoggerFactory
@@ -17,7 +17,7 @@ case class ValidateInputData(
 
   private val log = LoggerFactory.getLogger(classOf[ValidateInputData])
 
-  private implicit lazy val sqlContext = ModelMatrixAccess.hiveContext(sc)
+  private implicit lazy val sqlContext = ModelMatrix.hiveContext(sc)
 
   import com.collective.modelmatrix.cli.ASCIITableFormat._
   import com.collective.modelmatrix.cli.ASCIITableFormats._
