+++ modelmatrix-core/src/it/scala/com/collective/modelmatrix/FeaturizationSpec.scala
@@ -12,7 +12,7 @@ import scalaz.{-\/, \/, \/-}
 
 class FeaturizationSpec extends FlatSpec with GivenWhenThen with TestSparkContext {
 
-  val sqlContext = ModelMatrixAccess.sqlContext(sc)
+  val sqlContext = ModelMatrix.sqlContext(sc)
 
   val schema = StructType(Seq(
     StructField("auction_id", LongType),
