+++ modelmatrix-core/src/it/scala/com/collective/modelmatrix/transform/TransformerSpec.scala
@@ -1,15 +1,13 @@
 package com.collective.modelmatrix.transform
 
-import com.collective.modelmatrix.{ModelFeature, ModelMatrixAccess, TestSparkContext}
+import com.collective.modelmatrix.{ModelFeature, ModelMatrix, TestSparkContext}
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.types.{IntegerType, StringType, StructField, StructType}
 import org.scalatest.FlatSpec
 
-import scalaz.{\/-, -\/}
-
 class TransformerSpec extends FlatSpec with TestSparkContext {
 
-  val sqlContext = ModelMatrixAccess.sqlContext(sc)
+  val sqlContext = ModelMatrix.sqlContext(sc)
 
   val schema = StructType(Seq(
     StructField("adv_site", StringType),
