+++ modelmatrix-core/src/main/resources/db/migration/h2/mmc001__init_model_matrix_catalog_schema.sql
